Give hash points it's own file
